requirements.txt
firebase_admin
streamlit
streamlit_option_menu
python-dotenv
langchain
langchain-community
torch
accelerate
ctransformers[cuda]
transformers
sentence_transformers
streamlit_chat
faiss-cpu
altair
tiktoken
huggingface-hub
pypdf
#langchain-cli
ollama
llama-index-llms-ollama
langchain_ollama
